refactor: movidos scripts bash a la carpeta network_scripts , refactorizados los metodos de server
