+++ http/__init__.py
@@ -1,0 +1,1 @@
+# Package for HTTP handling (context & responder)
