+++ network_scripts/revocar.sh
@@ -10,21 +10,17 @@ CLIENT_IP="$1"
 CLIENT_MAC="$(echo "$2" | tr '[:upper:]' '[:lower:]')"
 LAN="wlp2s0"
 
-# -----------------------------
-# Detectar WAN automáticamente
-# -----------------------------
 WAN=$(ip route | grep default | awk '{print $5}' | head -n 1)
-echo "[+] WAN detectada automáticamente: $WAN"
+[ -z "$WAN" ] && echo "[!] No se detecta WAN" && exit 1
+
+echo "[+] Revocando IP $CLIENT_IP MAC $CLIENT_MAC (WAN $WAN)"
 
 iptables -D FORWARD -s "$CLIENT_IP" -o "$WAN" -m mac --mac-source "$CLIENT_MAC" -m state --state NEW,ESTABLISHED,RELATED -j ACCEPT 2>/dev/null || true
 iptables -D FORWARD -d "$CLIENT_IP" -i "$WAN" -m state --state ESTABLISHED,RELATED -j ACCEPT 2>/dev/null || true
-
 iptables -D FORWARD -s "$CLIENT_IP" -o "$WAN" -p icmp -j ACCEPT 2>/dev/null || true
 iptables -D FORWARD -d "$CLIENT_IP" -i "$WAN" -p icmp -j ACCEPT 2>/dev/null || true
-
 iptables -D FORWARD -s "$CLIENT_IP" -o "$WAN" -p udp --dport 53 -j ACCEPT 2>/dev/null || true
 iptables -D FORWARD -d "$CLIENT_IP" -i "$WAN" -p udp --sport 53 -j ACCEPT 2>/dev/null || true
-
 iptables -t nat -D PREROUTING -s "$CLIENT_IP" -p tcp --dport 80 -j RETURN 2>/dev/null || true
 
 echo "[+] Revocado: $CLIENT_IP ($CLIENT_MAC)"
